backend/pyproject.toml
[project]
name = "novellama"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "tiktoken>=0.9.0",
]

[project.scripts]
# Start flask server
novellama = "novellama:app.main"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

